fix(gigs): handle Supabase count errors on edit pages
Made-with: Cursor
